requirements.txt
discord
discord.py
requests
python-dotenv